made the delete question actually a question

+++ app/assets/javascripts/meals.js.coffee
@@ -7,12 +7,12 @@ meal.show_meal = (id) ->  document.location = '/meals/' + id
 
 meal.delete_meal = (id) ->
   global.askYesNoQuestion('Are you sure?', 'Do you really want to delete this meal?',
-    () ->
-      $.ajax({
-        url: '/meals/' + id,
-        type: 'delete',
-        success: () ->
-          document.location = '/meals'
-#          event.stopPropagation()
-      })
+    (res) ->
+      if (res == true)
+        $.ajax({
+          url: '/meals/' + id,
+          type: 'delete',
+          success: () ->
+            document.location = '/meals'
+        })
   )
